+++ src/Atata.WebDriverSetup/Http/HttpRequestExecutor.cs
@@ -12,6 +12,8 @@ namespace Atata.WebDriverSetup
 
         private readonly bool _checkCertificateRevocationList;
 
+        private readonly Action<HttpClientHandler> _httpClientHandlerConfigurationAction;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="HttpRequestExecutor"/> class.
         /// </summary>
@@ -21,10 +23,15 @@ namespace Atata.WebDriverSetup
         /// from the certificate store or if the caller is allowed to pass in a specific
         /// client certificate.
         /// </param>
-        public HttpRequestExecutor(IWebProxy proxy = null, bool checkCertificateRevocationList = true)
+        /// <param name="httpClientHandlerConfigurationAction">The configuration action of <see cref="HttpClientHandler"/>.</param>
+        public HttpRequestExecutor(
+            IWebProxy proxy = null,
+            bool checkCertificateRevocationList = true,
+            Action<HttpClientHandler> httpClientHandlerConfigurationAction = null)
         {
             _proxy = proxy;
             _checkCertificateRevocationList = checkCertificateRevocationList;
+            _httpClientHandlerConfigurationAction = httpClientHandlerConfigurationAction;
         }
 
         /// <inheritdoc/>
@@ -70,6 +77,8 @@ namespace Atata.WebDriverSetup
 
             if (_checkCertificateRevocationList)
                 httpClientHandler.CheckCertificateRevocationList = true;
+
+            _httpClientHandlerConfigurationAction?.Invoke(httpClientHandler);
 
             return new HttpClient(httpClientHandler, true);
         }